content-service/requirements.txt
aio-pika==9.5.7
asyncpg==0.30.0
fastapi==0.116.1
pydantic==2.11.7
pydantic-settings==2.10.1
PyJWT==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
SQLAlchemy==2.0.43
uvicorn==0.35.0
mutagen==1.47.0
slugify==0.0.1
boto3==1.40.76
